Created a short slideshow presentation for the first video, which is yet to be recorded. Started to break up the code into seperate tutorials, I got two of them out of the way 01 - Setting up the environment and getting a cube falling and hitting the ground visually 02 - Adding a sphere and cylinder to the scene
I did some things like make some nice textures in GIMP to make the visualisation a bit nicer. I went through and comment a lot of the code so a person can probably read it if they do not want to watch the videos.

I created a clean structure for each of the examples I created, this way the code is nice and seperated. Although I am going to try to stick with procedural functions that way people can derive their own objects from it all.